add  customer restful  action interface

+++ webService/src/main/java/com/uf/store/dao/mysql/ManagerRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.uf.store.dao.mysql.po.Manager;
 
-public interface ManagerRepository extends JpaRepository<Manager, Long> {
+public interface ManagerRepository extends JpaRepository<Manager, Integer> {
 	public Manager findTopByUserName(String userName);
 }
